Rename Rule Type sub-commands to `ruletype`
This even renamed the package. Cause, why not!?

Closes: #1755
